refactor: reorganize audio files into app/audio/ subdirectory
- Move 11 audio-related files from app/ to app/audio/
  - audio-context-manager.js
  - buffer-queue-node.js
  - codecs-browser.js
  - decode-worker.js
  - decoder-stream.js
  - encode-worker.js
  - encoder-stream.js
  - getusermedia.js
  - playback-buffer-processor.js
  - recorder-worker.js
  - voice.js

- Update import paths in app/index.js for audio modules
- Update require path in app/worker.js for codecs-browser
- Update smart-build.sh to copy AudioWorklet processors from app/audio/

This improves code organization by:
- Reducing app/ root from 19 to 8 JS files
- Grouping all audio processing logic in dedicated subdirectory
- Following same pattern as app/auth/ for better maintainability

Build verified successful (production + development modes).
